backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
python-magic>=0.4.27
python-magic-bin>=0.4.14
PyPDF2>=3.0.0
groq>=0.4.0
python-dateutil>=2.8.2
django-cors-headers>=4.3.0
